+++ auth_service/app/routers/auth.py
@@ -19,6 +19,13 @@ from app.dependencies.depend import (
     ensure_refresh_token_not_blacklisted,
     authentication_and_get_current_user
 )
+from app.utils import (
+    fetch_permissions_by_role_id,
+    fetch_role_name_by_role_id,
+    create_access_token,
+    create_refresh_token,
+    authenticate_user,
+)
 
 from env import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES, REFRESH_TOKEN_EXPIRE_DAYS
 
@@ -27,86 +34,7 @@ router = APIRouter(prefix="/auth", tags=["Auth"])
 bcrypt_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
 
-async def fetch_permissions_by_role_id(db: AsyncSession, role_id: int) -> list[str]:
-    result = await db.execute(
-        select(Permission.code)
-        .join(Role.permissions)
-        .where(Role.id == role_id)
-        .order_by(Permission.code)
-    )
-    return list(result.scalars().all())
 
-
-async def fetch_role_name_by_role_id(db: AsyncSession, role_id: int) -> str | None:
-    res = await db.execute(select(Role.name).where(Role.id == role_id))
-    return res.scalar_one_or_none()
-
-
-async def create_access_token(
-    username: str, user_id: int, role_id: int, expires_delta: timedelta
-):
-    permissions: list[str] = []
-    role_name: str | None = None
-    try:
-        async for db in get_db():
-            permissions = await fetch_permissions_by_role_id(db, role_id)
-            role_name = await fetch_role_name_by_role_id(db, role_id)
-            break
-    except Exception:
-        permissions = []
-        role_name = None
-
-    expire = datetime.utcnow() + expires_delta
-    payload = {
-        "sub": username,
-        "id": str(user_id),
-        "role_id": role_id,
-        "role_name": role_name,
-        "permissions": permissions,
-        "exp": int(expire.timestamp()),
-    }
-    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)
-
-
-async def create_refresh_token(
-    user_id: int, username: str, role_id: int, redis_client: redis.Redis
-):
-    role_name: str | None = None
-    try:
-        async for db in get_db():
-            role_name = await fetch_role_name_by_role_id(db, role_id)
-            break
-    except Exception:
-        role_name = None
-
-    expire = datetime.utcnow() + timedelta(days=REFRESH_TOKEN_EXPIRE_DAYS)
-    payload = {
-        "sub": username,
-        "id": str(user_id),
-        "role_id": role_id,
-        "role_name": role_name,  # <--- добавлено
-        "exp": int(expire.timestamp()),
-    }
-    token = jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)
-    await redis_client.setex(
-        f"refresh_{user_id}", REFRESH_TOKEN_EXPIRE_DAYS * 24 * 3600, token
-    )
-    return token
-
-
-async def authenticate_user(
-    db: Annotated[AsyncSession, Depends(get_db)],
-    username: str,
-    password: str,
-):
-    user = await db.scalar(select(User).where(User.name == username))
-    if not user or not bcrypt_context.verify(password, user.password_hash):
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid authentication credentials",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-    return user
 
 
 @router.post("/register", status_code=status.HTTP_201_CREATED)
@@ -131,7 +59,7 @@ async def register(db: Annotated[AsyncSession, Depends(get_db)], create_user: Cr
 @router.post("/login")
 async def login(
     db: Annotated[AsyncSession, Depends(get_db)],
-    form_data: Annotated[OAuth2PasswordRequestForm, Depends()],  # форма username/password
+    form_data: Annotated[OAuth2PasswordRequestForm, Depends()],
     redis_client: redis.Redis = Depends(get_redis),
 ):
     user = await authenticate_user(db, form_data.username, form_data.password)
